pyproject.toml
[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
    # isort
    "I",
]

[tool.poetry]
name = "quiz-app"
version = "0.1.0"
description = "Backend for Quiz App"
license = "MIT"
authors = ["Matias B"]
readme = "README.md"
repository = "https://github.com/mbrignone/quiz-app"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.0.3"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
pytest = "^7.4.3"
poethepoet = "^0.24.1"
pre-commit = "^3.5.0"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
start = "python app/main.py"
test = "pytest --version"
lint = "ruff check ."
format = "ruff format ."
precommit = "pre-commit run --all-files"
